Update Codegen.Models to handle $ref to non-models
Codegen.Models previously just ignored any top-level models that were
not actually models, but datatypes intended to be used with $ref.  This
meant that we'd get decode errors when decoding datatypes that relied on
these $refs.

This commit fixes that: we now copy the fields from these refs in place
when building our ModelDescs & PropertyDescs.