Add Codegen.Apis & Kazan.Apis generated from it.
This adds the code generation system for the various API namespaces &
functions, and uses them to generate the Kazan.Apis namespace.  This
namespace contains a sub-namespace for each tag that is provided in the
kubeOAI spec.  Within those namespaces are functions that can be used to
generate requests.

Those requests can be fed into the clients that will be coming in a
later commit.