Document type info for non-model model properties
Before we didn't print out any information in the docs if a models property was
some basic type like a string or a number.  Initially I was intending for this
to be done via an autogenerated type-spec, but for now it seems a good idea to
have this in there as well.