Name generated model modules more concicsely.
Kubernetes updated the structure of their open API specs in 1.6.  This
made the generated module names in kazan extremely verbose.

This updates the Codegen.Models module to strip off some prefixes from
these names.  A mapping of the name changes can be found in
`strip_module_name_prefixes` in this commit.