Make module generating names safe to call.
A couple of the functions that are involved in the codegen process were
constructing module names with `Module.concat`.  This is fine at build
time, because we _need_ to create new atoms.

However, I want to expose these functions to users, potentially to call
with unverified user input (and I suspect some of them were called
already when decoding structs from kube).  So, I've modified them to
take an `unsafe` option.  This will be set when doing codegen, but unset
elsewhere.  When this is not set we use `Module.safe_concat` - which
will not create any new atoms.