Support for specifying auth directly in mix config (#41)
The documentation for Kazan says that we support specifying a server map
in the mix config.  This was true, but the auth map within the config
was not being converted to one of the `Kazan.Server.*` structs. This
meant that the `Kazan.run` calls would fail as they pattern match on the
struct type.

This updates `Kazan.Server` with a function that properly converts maps
to the appropriate nested structs, which should allow people to specify
the auth details in their mix config.

Fixes #39