Add typespecs to generated API functions.
This updates the API codegen module to add typespecs to the functions it
generates.  This improves documentation and should improve the
experience for users using dialyzer to analyse their code.

At the moment we don't have typespecs for the models, so these are just
marked as structs.  This should be changed in a future commit.