Handle 410 Expired messages without crashing. (#76)
I have observed messages with code `410` and reason `Expired` arriving and causing the `Kazan.Watcher` process to crash with a function clause error like follows:

```
2020-03-30 14:26:51.276 [error] pid=<0.1720.0> module=gen_server function=error_info/7 line=889  GenServer #PID<0.1720.0> terminating
** (FunctionClauseError) no function clause matching in Kazan.Watcher.extract_rv/1
    (kazan 0.11.0) lib/kazan/watcher.ex:309: Kazan.Watcher.extract_rv(%{"apiVersion" => "v1", "code" => 410, "kind" => "Status", "message" => "The resourceVersion for the provided watch is too old.", "metadata" => %{}, "reason" => "Expired", "status" => "Failure"})
    (kazan 0.11.0) lib/kazan/watcher.ex:281: Kazan.Watcher.process_line/3
    (elixir 1.10.1) lib/enum.ex:2111: Enum."-reduce/3-lists^foldl/2-0-"/3
    (kazan 0.11.0) lib/kazan/watcher.ex:201: Kazan.Watcher.handle_info/2
    (stdlib 3.11) gen_server.erl:637: :gen_server.try_dispatch/4
    (stdlib 3.11) gen_server.erl:711: :gen_server.handle_msg/6
    (stdlib 3.11) proc_lib.erl:249: :proc_lib.init_p_do_apply/3
Last message: %HTTPoison.AsyncChunk{chunk: "{\"type\":\"ERROR\",\"object\":{\"kind\":\"Status\",\"apiVersion\":\"v1\",\"metadata\":{},\"status\":\"Failure\",\"message\":\"The resourceVersion for the provided watch is too old.\",\"reason\":\"Expired\",\"code\":410}}\n", id: #Reference<0.1835110855.1638924289.183387>}
```

This intention of this PR is to handle this scenario more gracefully, in the same way as the `410 Gone` message.

The Kubernetes documentation about this `410 Expired` message is very limited. I'm curious to hear if someone knows why this may happen (in my case around every 10 minutes) and if there is a way to avoid this, since it seems to me that there might be a risk of losing events when this happens.

Any comments and suggestions are appreciated.